Split class file.
