+++ Assets/Script/FishManager.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 42441b123c80df04486962f8e8fee7f7
+timeCreated: 1503227275
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
